Merged PR 10: Database file in repo, changes to contract validation
Since we can't create an SQLite database file on Azure, we need to have it premade. I've added an empty file with correct name.
Also contract validation was changed, because previously it checked if the file name is correct, and I think it may be problematic to preserve the original file name when it is downloaded by comparer.